+++ addOns/automation/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Adjust columns in plan display after loading a plan.
+- Include cause of invalid URL in error message.
 
 ### Removed
 - The Params automation support was moved into the Params add-on (Issue 9210).