+++ addOns/spider/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Maintenance changes.
+- Include cause of invalid URL in error message of Spider job.
 
 ## [0.20.0] - 2026-04-02
 ### Added